Moved original batch code back to ingester.go
